Native macOS GUI, fidelity tests, three codec bug fixes, cleanup
The Tk GUI showed an empty window: the only Tk on stock macOS is Apple's
8.5.9 from 2010, which crashes during widget construction on current macOS.
Rewrote gui.py on native dialogs via osascript -- no Tk, no extra install.

Added test_fast.py (29 cases, run through both the C path and the numpy
fallback). It immediately found three real bugs, now fixed:

  * the varint tail was 32-bit and silently truncated any value past 2^31,
    well inside the int64 range the codec accepts. Tail width is now chosen
    per array, 4 or 8 bytes, so correctness costs nothing on real data.
  * text was stored newline-joined, so any cell containing a newline was
    split in two. Groups now carry explicit lengths, but only when a value
    actually contains a newline -- the joined layout xz likes is kept.
  * an empty table divided by zero in the parent-selection entropy.

Also: numpy-fallback zigzag used arithmetic that overflows int64 on
differences near the +/-2^62 limit; switched to the shift/xor form.

Removed the functional-dependency exploration scripts, which that line of
work no longer needs. Added bench.py for size and speed in both directions.

All five data.gov tables unchanged: 1.34x-2.01x over xz -9e, all exact.

+++ tcz.c
@@ -163,9 +163,11 @@ int64_t fmt_fixed(const int64_t *in, int64_t n, int32_t dec,
 /* ------------------------------------------------------------- varints */
 
 /* Zigzag each value; small ones take a byte, the rest an escape plus a
- * 32-bit tail entry. Returns how many went to the tail. */
+ * 64-bit tail entry. The tail must be 64 bits: a 32-bit tail silently
+ * truncated any value past 2^31, which is well inside the int64 range this
+ * codec accepts. */
 int64_t pack_ints(const int64_t *in, int64_t n,
-                  uint8_t *head, uint32_t *tail)
+                  uint8_t *head, uint64_t *tail)
 {
     int64_t nbig = 0;
     for (int64_t i = 0; i < n; i++) {
@@ -175,14 +177,14 @@ int64_t pack_ints(const int64_t *in, int64_t n,
             head[i] = (uint8_t)u;
         } else {
             head[i] = ESCAPE;
-            tail[nbig++] = (uint32_t)u;
+            tail[nbig++] = u;
         }
     }
     return nbig;
 }
 
 void unpack_ints(const uint8_t *head, int64_t n,
-                 const uint32_t *tail, int64_t *out)
+                 const uint64_t *tail, int64_t *out)
 {
     int64_t nbig = 0;
     for (int64_t i = 0; i < n; i++) {
